app/requirements.txt
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.4
django-cors-headers==2.0.2
django-pandas==0.4.1
djangorestframework==3.6.2
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
pandas==0.19.2
psycopg2==2.7.1
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
pyodbc
django-pyodbc-azure
gunicorn==19.3.0
django-allauth>=0.24.1
django-rest-swagger==2.0.7
django-rest-auth==0.9.1
appdirs==1.4.3
bleach==2.0.0
colorama==0.3.7
decorator==4.0.11
django-extensions==1.7.8
django-mssql==1.8
drfdocs==0.0.11
entrypoints==0.2.2
html5lib==0.999999999
numpy
packaging==16.8
pandocfilters==1.4.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
pyparsing==2.2.0
pyzmq==16.0.2
qtconsole==4.3.0
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
testpath==0.3
tornado==4.4.3
